+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **FIX**: some `G.FDR` configurations caused incorrect rendering in grid visualizer
 - **NEW**: `PROB` accepts an `ELSE` condition
 - **NEW**: alias `PRB` for `PROB`
+- **NEW**: new op: `WTOSS`, alias `WT`
 
 ## v4.0.0